Adding the possibility to disable the PlayerInteractEvent use for the LimitedItemInteract + refactoring

+++ src/main/java/com/gamingmesh/jobs/commands/list/leave.java
@@ -40,7 +40,7 @@ public class leave implements Cmd {
 		Util.LEAVECONFIRM.add(uuid);
 
 		plugin.getServer().getScheduler().runTaskLater(plugin, () -> Util.LEAVECONFIRM.remove(uuid),
-		    (long) (20 * Jobs.getGCManager().ConfirmExpiryTime));
+				20 * Jobs.getGCManager().ConfirmExpiryTime);
 
 		pSender.sendMessage(Jobs.getLanguage().getMessage("command.leave.confirmationNeed", "[jobname]",
 		    job.getJobDisplayName(), "[time]", Jobs.getGCManager().ConfirmExpiryTime));
